change dtype of label to x.dtype instead of torch.float in image_with_label.py

+++ models/models/mmclassifiers/image_with_label.py
@@ -36,7 +36,7 @@ class ImageWithLabelClassifier(ImageClassifier):
         if isinstance(x, tuple):
             x = x[-1]
         x = self.img_fc(x)
-        label = self.label_fc(label.to(torch.float))
+        label = self.label_fc(label.to(x.dtype))
         x = torch.cat((x, label), dim = 1)
 
         losses = dict()
@@ -52,7 +52,7 @@ class ImageWithLabelClassifier(ImageClassifier):
         if isinstance(x, tuple):
             x = x[-1]
         x = self.img_fc(x)
-        label = self.label_fc(label.to(torch.float))
+        label = self.label_fc(label.to(x.dtype))
         x = torch.cat((x, label), dim = 1)
 
         res = self.head.simple_test(x, **kwargs)
